pyproject.toml
[tool.poetry]
name = "TrafficToll"
version = "0.1.0"
description = ""
authors = ["cryzed <cryzed@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
psutil = "^5.5"
"ruamel.yaml" = "^0.15.87"
loguru = "^0.2.5"
PyQt5 = "^5.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
